pyproject.toml
[project]
name = "ggj-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=24.1.1",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "waitress>=3.0.2",
]
